A new beginning
On a random day, all the civilizations joined forces to leave the dying state of the world in which they are inhabiting. They left in fleets of much grandeur and it was at that time when a vortex opened somewhere between the stars and they entered it where they had the encounter with the Beings. They acquired exotic technologies from them but little did they know that it was all part of a much greater story and that there was nothing random about that day.
The Exotic technologies include:
Bombs and Nukes
Guided missile systems
Power manipulation devices
Unconventional weapon systems
and much more.